pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "ycleptic"
version = "2.0.1"
authors = [
  { name="Cameron F Abrams", email="cfa22@drexel.edu" },
]
description = "A YAML-format configuration generator"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml>=6"
]
[project.scripts]
yclept = "ycleptic.cli:cli"
[project.urls]
"Source" = "https://github.com/cameronabrams/ycleptic"
"Documentation" = "https://ycleptic.readthedocs.io/en/latest/"
"Bug Tracker" = "https://github.com/cameronabrams/ycleptic/issues"
# [tool.pytest.ini_options]
# log_cli = true